rename Strategies to strategies
