+++ src/strategies/index.ts
