+++ src/strategies/listVerticalStrategy.ts
