+++ src/strategies/singleObjectStrategy.ts
